requirements.txt
telethon>=1.36.0
python-dotenv>=1.0.1
PyYAML>=6.0
requests>=2.28.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
